@@ -6,39 +6,38 @@
   "bugs": "https://github.com/oclif/plugin-plugins/issues",
   "dependencies": {
     "@oclif/color": "^0.0.0",
-    "@oclif/command": "^1.5.0",
+    "@oclif/command": "^1.5.1",
     "chalk": "^2.4.1",
     "cli-ux": "^4.8.2",
     "debug": "^4.0.1",
     "fs-extra": "^7.0.0",
     "http-call": "^5.2.2",
-    "load-json-file": "^5.0.0",
+    "load-json-file": "^5.1.0",
     "npm-run-path": "^2.0.2",
     "semver": "^5.5.1",
     "tslib": "^1.9.3",
-    "yarn": "^1.9.4"
+    "yarn": "^1.10.1"
   },
   "devDependencies": {
     "@oclif/config": "^1.7.6",
-    "@oclif/dev-cli": "^1.18.0",
-    "@oclif/errors": "^1.2.0",
+    "@oclif/dev-cli": "^1.18.1",
+    "@oclif/errors": "^1.2.1",
     "@oclif/plugin-help": "^2.1.2",
     "@oclif/test": "^1.2.1",
     "@oclif/tslint": "^3.1.0",
-    "@types/chai": "^4.1.4",
+    "@types/chai": "^4.1.6",
     "@types/fs-extra": "^5.0.4",
-    "@types/load-json-file": "^2.0.7",
     "@types/mocha": "^5.2.5",
-    "@types/node": "^10.9.4",
+    "@types/node": "^10.11.3",
     "@types/semver": "^5.5.0",
     "@types/supports-color": "^5.3.0",
-    "chai": "^4.1.2",
+    "chai": "^4.2.0",
     "fancy-test": "^1.4.1",
     "globby": "^8.0.1",
     "mocha": "^5.2.0",
     "ts-node": "^7.0.1",
     "tslint": "^5.11.0",
-    "typescript": "^3.0.3"
+    "typescript": "^3.1.1"
   },
   "engines": {
     "node": ">=8.0.0"
